Refactors controllers
